backend/requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Vision AI
google-generativeai==0.3.1

# Image Processing
Pillow==10.1.0

# Environment Management
python-dotenv==1.0.0
